+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add `AGENT.md` and configure symlink to `.github/copilot-instructions.md`
 - Configure `shadcn/ui` and configure fonts
 - Add some standard `shadcn/ui` components
+- Add some common components, including `Typography`, `List` and `Link`
 - Add the `AppShell` component, which includes a sidebar and main content area
 - Implemented a mock homepage
 - Configured unit testing with React Testing Library, added sample unit test